requirements.txt
pillow
joblib
torch 
torchvision
fastai
numpy==1.26.4